requirements.txt
pandas>=1.5.0,<4.0.0
numpy>=1.23.0
requests>=2.28.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
pykrx>=1.2.0
openai>=2.0.0
groq>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0
flask>=3.0.0
flask-cors>=4.0.0
